requirements.txt
pycountry<=22.3.5
cerberus<=1.3.4      
xmltodict<=0.13.0      
dicttoxml<=1.7.4 
multipledispatch<=0.6.0     
pycryptodome<=3.14.1    
mysql    
pymysql<=1.0.2                
mysql-connector-python<=8.0.28
sshtunnel<=0.4.0
rentdynamics<=1.0.1
protobuf==3.19.4
place-api==0.6.0
pydantic
SQLAlchemy-Paginator==0.2
psycopg2==2.7.4
zeep
python-dateutil
jira
